\name{GenomicFiles-deprecated}
\alias{GenomicFiles-deprecated}
\alias{getVCFPath}

\title{Deprecated functions in package \sQuote{GenomicFiles}}

\description{
  These functions are provided for compatibility with older versions
  of \sQuote{GenomicFiles} only, and will be defunct at the next release.
}

\details{

  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{

    \item{getVCFPath(vs, chrtok): \code{files(vs)[chrtok]}}

  }
}
